Moving warming level util functions to util/warming_levels.py instead of util/utils.py
